*: decouple HTTP mutator and fix code structure again
